refactor: move help message outside navigation and add structure
