+++ src/components/HelpMessage/index.ts
@@ -1,0 +1,3 @@
+import HelpMessageContainer from './HelpMessageContainer';
+
+export default HelpMessageContainer;
